+++ internal/sampling/grpc/grpc_handler.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2018 The Jaeger Authors.
 // SPDX-License-Identifier: Apache-2.0
 
-package sampling
+package grpc
 
 import (
 	"context"
@@ -10,19 +10,19 @@ import (
 	"github.com/jaegertracing/jaeger/proto-gen/api_v2"
 )
 
-// GRPCHandler is sampling strategy handler for gRPC.
-type GRPCHandler struct {
+// Handler is sampling strategy handler for gRPC.
+type Handler struct {
 	samplingProvider samplingstrategy.Provider
 }
 
-// NewGRPCHandler creates a handler that controls sampling strategies for services.
-func NewGRPCHandler(provider samplingstrategy.Provider) GRPCHandler {
-	return GRPCHandler{
+// NewHandler creates a handler that controls sampling strategies for services.
+func NewHandler(provider samplingstrategy.Provider) Handler {
+	return Handler{
 		samplingProvider: provider,
 	}
 }
 
 // GetSamplingStrategy returns sampling decision from store.
-func (s GRPCHandler) GetSamplingStrategy(ctx context.Context, param *api_v2.SamplingStrategyParameters) (*api_v2.SamplingStrategyResponse, error) {
+func (s Handler) GetSamplingStrategy(ctx context.Context, param *api_v2.SamplingStrategyParameters) (*api_v2.SamplingStrategyResponse, error) {
 	return s.samplingProvider.GetSamplingStrategy(ctx, param.GetServiceName())
 }
